Rename subcommand classes for announcement management
