+++ toolkitplugin/src/main/java/org/fourz/rvnktools/announceManager/command/AnnounceSubCommandHelp.java
@@ -1,5 +1,5 @@
 
-package org.fourz.rvnktools.announceManager.subcommand;
+package org.fourz.rvnktools.announceManager.command;
 
 import org.bukkit.entity.Player;
 import org.fourz.rvnktools.RVNKTools;
